# database: speed up UpdateRepoSizes