--- a/internal/database/gitserver_repos_test.go
+++ b/internal/database/gitserver_repos_test.go
@@ -1048,7 +1048,7 @@ func TestGitserverUpdateRepoSizes(t *testing.T) {
 		repo2.Name: 500,
 		repo3.Name: 800,
 	}
-	numUpdated, err := db.GitserverRepos().UpdateRepoSizes(ctx, shardID, sizes)
+	numUpdated, err := db.GitserverRepos().UpdateRepoSizes(ctx, logger, shardID, sizes)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1082,7 +1082,7 @@ func TestGitserverUpdateRepoSizes(t *testing.T) {
 	}
 
 	// update again to make sure they're not updated again
-	numUpdated, err = db.GitserverRepos().UpdateRepoSizes(ctx, shardID, sizes)
+	numUpdated, err = db.GitserverRepos().UpdateRepoSizes(ctx, logger, shardID, sizes)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1094,7 +1094,7 @@ func TestGitserverUpdateRepoSizes(t *testing.T) {
 	sizes = map[api.RepoName]int64{
 		repo3.Name: 900,
 	}
-	numUpdated, err = db.GitserverRepos().UpdateRepoSizes(ctx, shardID, sizes)
+	numUpdated, err = db.GitserverRepos().UpdateRepoSizes(ctx, logger, shardID, sizes)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1111,13 +1111,51 @@ func TestGitserverUpdateRepoSizes(t *testing.T) {
 			repo3.Name: 789 + batchSize,
 		}
 
-		numUpdated, err = gitserverRepoStore.updateRepoSizesWithBatchSize(ctx, sizes, int(batchSize))
+		numUpdated, err = gitserverRepoStore.updateRepoSizesWithBatchSize(ctx, logger, sizes, int(batchSize))
 		if err != nil {
 			t.Fatal(err)
 		}
 		if have, want := numUpdated, 3; have != want {
 			t.Fatalf("wrong number of repos updated. have=%d, want=%d", have, want)
 		}
+	}
+}
+
+func TestGitserverUpdateRepoSizes_LargeAmountOfRepos(t *testing.T) {
+	if testing.Short() {
+		t.Skip()
+	}
+
+	logger := logtest.Scoped(t)
+	db := NewDB(logger, dbtest.NewDB(t))
+	ctx := context.Background()
+
+	// Large number of repositories.
+	// NOTE: Tweak this to ensure the code runs fast. When you modify the code,
+	// you should change this 20k or 60k.
+	const count = 6_000
+	namesToSize := make(map[api.RepoName]int64, count)
+
+	reposBatch := make([]*types.Repo, 0, 1000)
+	for i := 0; i < count; i++ {
+		name := fmt.Sprintf("github.com/sourcegraph/repo-%d", i)
+		r := &types.Repo{Name: api.RepoName(name)}
+		namesToSize[r.Name] = int64(i + 1)
+
+		reposBatch = append(reposBatch, r)
+		if i%5000 == 0 || i == count-1 {
+			createTestRepos(ctx, t, db, types.Repos(reposBatch))
+			reposBatch = reposBatch[0:0]
+			t.Logf("Creating %d repos. %d to go", i, count-i)
+		}
+	}
+
+	numUpdated, err := db.GitserverRepos().UpdateRepoSizes(ctx, logger, shardID, namesToSize)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if have, want := numUpdated, len(namesToSize); have != want {
+		t.Fatalf("wrong number of repos updated. have=%d, want=%d", have, want)
 	}
 }