Noticed as part of #wg-cheaper-sourcegraph that the calls to UpdateRepoSizes on sourcegraph.com would never, ever return. Instead the goroutine would get stuck, somehow connection between client and DB would get lost, DB would think client is still there, and the whole thing might have locked rows, and it all became a big disaster.

Digging into this I found that even locally, I could never get the test to update, say, 100k repos with the given batch size. It would time out after 10min.

So what I was to

1. Remove the transaction and the locking -- we update two fields anywhere, that aren't updated by anything else (and in the case of `updated_at` it doesn't matter)
2. Reduce the batch size drastically
3. Add a test to play around with large repo sizes
4. Add a logger

When running the test with 20k repos, I noticed the following:

    Batch size 100: test takes 35s
    Batch size 1000: test takes 64s

Then, with 5k repos, I played with the batch size some more and found:

    Batch size 100: one batch takes ~50ms to update.
    Batch size 500: one batch takes ~650ms to update.

So let's go with 100.

On sourcegraph.com, we right now have 770k repos per shard. That means we end up with 7.700 batches. That should take ~6-7min.

That's still better than one continuous query that never finishes.


## Test plan

- Unit test